requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
opencv-python-headless
face-recognition
dlib-bin  # ใช้เวอร์ชันที่ติดตั้งได้ง่ายกว่า
